chore: clean up ruff import rules
