+++ src/dify_plugin/_gevent.py
@@ -1,0 +1,4 @@
+from gevent import monkey
+
+# Keep the package-level gevent patch as the first import side effect.
+monkey.patch_all(sys=True)
